theme setting done + use different viewModel
